Break circular dependency: init -> * -> init by extracting shutdown.h
>>> based on bitcoin@1fabd59e7e870fae73bfbcfb227dd7452de94726
