+++ src/pivxd.cpp
@@ -15,6 +15,7 @@
 #include "init.h"
 #include "masternodeconfig.h"
 #include "noui.h"
+#include "shutdown.h"
 #include "util/system.h"
 
 #include <stdio.h>
