+++ src/qt/winshutdownmonitor.cpp
@@ -6,7 +6,7 @@
 #include "winshutdownmonitor.h"
 
 #if defined(Q_OS_WIN)
-#include "init.h"
+#include "shutdown.h"
 
 #include <windows.h>
