+++ src/rpc/server.cpp
@@ -8,9 +8,9 @@
 #include "rpc/server.h"
 
 #include "fs.h"
-#include "init.h"
 #include "key_io.h"
 #include "random.h"
+#include "shutdown.h"
 #include "sync.h"
 #include "guiinterface.h"
 #include "util/system.h"
