+++ src/shutdown.cpp
@@ -1,0 +1,25 @@
+// Copyright (c) 2009-2010 Satoshi Nakamoto
+// Copyright (c) 2009-2018 The Bitcoin Core developers
+// Copyright (c) 2021 The PIVX developers
+// Distributed under the MIT software license, see the accompanying
+// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+
+#include "shutdown.h"
+
+#include <atomic>
+
+
+static std::atomic<bool> fRequestShutdown{false};
+
+void StartShutdown()
+{
+    fRequestShutdown = true;
+}
+void AbortShutdown()
+{
+    fRequestShutdown = false;
+}
+bool ShutdownRequested()
+{
+    return fRequestShutdown;
+}
