+++ src/validation.cpp
@@ -25,7 +25,6 @@
 #include "evo/specialtx.h"
 #include "flatfile.h"
 #include "guiinterface.h"
-#include "init.h"
 #include "invalid.h"
 #include "interfaces/handler.h"
 #include "legacy/validation_zerocoin_legacy.h"
@@ -37,6 +36,7 @@
 #include "pow.h"
 #include "reverse_iterate.h"
 #include "script/sigcache.h"
+#include "shutdown.h"
 #include "spork.h"
 #include "sporkdb.h"
 #include "evo/evodb.h"
