+++ test/lint/lint-circular-dependencies.sh
@@ -25,9 +25,6 @@ EXPECTED_CIRCULAR_DEPENDENCIES=(
     "evo/deterministicmns -> evo/specialtx -> evo/deterministicmns"
     "evo/deterministicmns -> masternodeman -> evo/deterministicmns"
     "evo/providertx -> evo/specialtx -> evo/providertx"
-    "init -> masternodeman -> init"
-    "init -> rpc/server -> init"
-    "init -> validation -> init"
     "kernel -> validation -> kernel"
     "masternode -> masternode-sync -> masternode"
     "masternode -> masternodeman -> masternode"
